imho opinion a better handling of the names
Material name will be object name as a default. When the create button
is pressed twice, a name object.001 is automatically created.

To save this, I added a new filepath to the ExportHelper which
predefines the name.

Some tests must still be done in connection with
makeclothes and already existent files (textures), especially
when material is copied in same folder. (e.g. 3 mhmat files, 3 diffuse
textures but only one common normalmap)