changed the texture node tests (filenames)
I added an extra test isfile. So the copy program will not throw an
exception later. Furthermore I forward the statements to report command
(pls. no prints in a logfile if sth. is obvious ;-) )

The test will be added in makeclothes also. But therefore I have to
change the write command so it does not work differently if you call in
direct in makeskin or in makeclothes ...